// src/enums/ColumnsCategories.ts
export enum ColumnsCategories {
    //Folder = 'FolderPath',
    Stage = 'Stage',
    Phase = 'Phase',
    //Category = 'Kategorie',
    Description = 'FolderDescription',
    DescriptionEn = 'FolderDescription_en',
    StageEn = 'Stage_en',
    PhaseEn = 'Phase_en',
    CategoryEn = 'Category_en',
    RequiredDocuments = 'RequiredDocuments',
    //Deadline = 'Deadline',
    Order = 'CategoryOrder',
    CategoryStatus = 'CategoryStatus'
}